make all code pretty

+++ childrenevents/controller/sport_event_controller.py
@@ -4,7 +4,7 @@ import json
 from flask import Response, request
 from childrenevents.model.sport_event import SportEvent
 from flask_restful import Resource
-from childrenevents.datebase.db import ma, db
+from childrenevents.app import ma, db
 
 
 class SportEventSchema(ma.Schema):
@@ -12,9 +12,7 @@ class SportEventSchema(ma.Schema):
         # Fields to expose
         fields = (
             'id', 'name', 'contact_number', 'price_in_uah', 'max_quantity_of_children', 'duration_in_minutes', 'venue',
-            'location',
-            'sport_equipment'
-        )
+            'location', 'sport_equipment')
 
 
 sport_event_schema = SportEventSchema()
@@ -29,18 +27,8 @@ class SportEventsApi(Resource):
         return Response(sport_events, mimetype="application/json", status=200)
 
     def post(self):
-        name = request.json['name']
-        contact_number = request.json['contact_number']
-        price_in_uah = request.json['price_in_uah']
-        max_quantity_of_children = request.json['max_quantity_of_children']
-        duration_in_minutes = request.json['duration_in_minutes']
-        venue = request.json['venue']
-        location = request.json['location']
-        sport_equipment = request.json['sport_equipment']
-
-        new_sport_event = SportEvent(name, contact_number, price_in_uah, max_quantity_of_children, duration_in_minutes,
-                                     venue, location, sport_equipment)
-
+        body = request.get_json()
+        new_sport_event = SportEvent(**body)
         db.session.add(new_sport_event)
         db.session.commit()
         return sport_event_schema.jsonify(new_sport_event)
@@ -48,36 +36,26 @@ class SportEventsApi(Resource):
 
 class SportEventApi(Resource):
     def put(self, id):
-        sport_event = SportEvent.query.get(id)
-        name = request.json['name']
-        contact_number = request.json['contact_number']
-        price_in_uah = request.json['price_in_uah']
-        max_quantity_of_children = request.json['max_quantity_of_children']
-        duration_in_minutes = request.json['duration_in_minutes']
-        venue = request.json['venue']
-        location = request.json['location']
-        sport_equipment = request.json['sport_equipment']
+        body = request.get_json()
+        sport_event: SportEvent = SportEvent.query.get(id)
+        print(str(sport_event).center(100, ' '))
+        if sport_event is None:
+            return Response(status=404)
         old_sport_event = copy.deepcopy(sport_event)
-
-        sport_event.name = name
-        sport_event.contact_number = contact_number
-        sport_event.price_in_uah = price_in_uah
-        sport_event.max_quantity_of_children = max_quantity_of_children
-        sport_event.duration_in_minutes = duration_in_minutes
-        sport_event.venue = venue
-        sport_event.location = location
-
-        sport_event.sport_equipment = sport_equipment
-
+        sport_event.copy_params(**body)
         db.session.commit()
         return sport_event_schema.jsonify(old_sport_event)
 
     def delete(self, id):
         sport_event = SportEvent.query.get(id)
+        if sport_event is None:
+            return Response(status=404)
         db.session.delete(sport_event)
         db.session.commit()
-        return '', 200
+        return Response(status=200)
 
     def get(self, id):
         sport_event = SportEvent.query.get(id)
+        if sport_event is None:
+            return Response(status=404)
         return sport_event_schema.jsonify(sport_event)
